+++ modules/plugins/interrupts.lua
@@ -14,9 +14,9 @@ local CHANNELS = {
 }
 
 local function Interrupt_OnEvent(self, event)
-    local timestamp, eventType, hideCaster, sourceGUID, sourceName, sourceFlags,
-        sourceRaidFlags, destGUID, destName, destFlags, destRaidFlags, spellID,
-        spellName, spellSchool = CombatLogGetCurrentEventInfo()
+    local timestamp, eventType, hideCaster, sourceGUID, sourceName,
+    sourceFlags, sourceRaidFlags, destGUID, destName, destFlags,
+    destRaidFlags = CombatLogGetCurrentEventInfo()
 
     local chatType
     if (IsInGroup(LE_PARTY_CATEGORY_INSTANCE)) then
@@ -31,14 +31,20 @@ local function Interrupt_OnEvent(self, event)
 
     -- check if event is interrupt
     if ((eventType == "SPELL_INTERRUPT") and (sourceGUID == UnitGUID("player"))) then
+        local spellID, spellName, spellSchool,
+        extraSpellID, extraSpellName, extraSchool = select(12, CombatLogGetCurrentEventInfo())
+
         -- check if channel is enable for announcing
         if (not CHANNELS[chatType]) then
-            -- check if SAY channel is disable
-            if (chatType == "SAY") then return end
             chatType = "SAY"
         end
 
-        SendChatMessage("Interrupted " .. destName .. " " .. GetSpellLink(spellID) .. "!", chatType)
+        if (C.Interrupts.SpellLink) then
+            local extraSpellLink = GetSpellLink(extraSpellID)
+            SendChatMessage("Interrupted " .. destName .. " " .. extraSpellLink .. "!", chatType)
+        else
+            SendChatMessage("Interrupted " .. destName .. " " .. extraSpellName .. "!", chatType)
+        end
     end
 end